Get CI back to green
CI has been red since 2026-07-30 (four consecutive runs). The blocking Go lint
step failed first, which masked three further gates behind it — each only
visible once the one before it passed.

Go lint (curated), 25 issues:
  - 11 lll: long function signatures, wrapped properly rather than suppressed.
  - 8 funlen, 2 gocognit, 1 gocyclo: SQL-heavy repository functions and
    branch-per-field validators, given //nolint with a reason, matching the 136
    existing directives in this module. Splitting a function that is one large
    query plus its row scan hides the query from its mapping.
  - 1 gofmt: mine. Adding AffiliateSignup to UpdateMarketingFlagsCommand
    changed the struct's alignment column and I did not re-run gofmt.
  - 1 unconvert: int64(chargeMinor) where chargeMinor is already int64.
  - 1 gosec G705 "XSS via taint": a false positive, but not suppressed blindly.
    The CSV export is served as text/csv with Content-Disposition: attachment,
    and every cell is machine-generated (RFC3339 timestamp, status/type enum,
    strconv integer) so CSV formula injection has no way in either. Added
    X-Content-Type-Options: nosniff, which closes the one real residual risk —
    a browser sniffing the body as HTML — then a scoped nolint explaining why.

eslint:
  - no-useless-catch in load-portal.server.ts: mine. When withAffiliateAuth took
    over 401 handling I left a catch that only rethrows. Removed; the comment
    explaining why errors must propagate stays.
  - Two unused eslint-disable directives, one of them made unused by the
    duplicate-property cleanup lowering DashboardWorkspace's complexity.
  - Size/complexity disables on three large presentational components.
  - .artifacts and .raven added to the eslint ignores: gitignored, so CI never
    saw them, but they failed `pnpm check` locally. Local and CI disagreeing
    about whether the tree is clean is its own bug.

check-file-size: six affiliate files over budget, allowlisted with the concrete
split queued for each. This list is meant to shrink, so they are recorded as
debt, not blessed — they merged while CI was already red and were never gated.

Verified locally against every CI step: golangci-lint 0 issues, go build/vet
clean, full go test suite passes, pnpm check passes, pnpm test passes.

+++ apps/api/internal/application/auth/service_subscription.go
@@ -276,7 +276,7 @@ func (s Service) InitializeSubscriptionAuthorization(
 				BusinessID:       subscription.BusinessID,
 				SubscriptionID:   subscription.SubscriptionID,
 				PaymentReference: checkoutRef,
-				GrossMinor:       int64(chargeMinor),
+				GrossMinor:       chargeMinor,
 			},
 		)
 		switch {
